Auswertung und Median gefixt ConsistencyQuestion in PollController eingebaut Random-Antworten-Funktion erstellt Landingpage mit Polls gefüttert

+++ src/main/java/gpse/umfrato/domain/cmd/QuestionCmd.java
@@ -13,11 +13,11 @@ public class QuestionCmd {
 
     private long questionId;
 
-    private float endValue;
+    private double endValue;
 
-    private float startValue;
+    private double startValue;
 
-    private float stepSize;
+    private double stepSize;
 
     private String belowMessage;
